+++ Assets/Scripts/Map/Tiles/MappedTiles/Buildings/ConnectedTile.cs
@@ -71,13 +71,15 @@ public class ConnectedTile : PoweredBuildingTile
 		}
 	}
 
-	public override void Show(bool isShown)
+	public override bool Show(bool isShown)
 	{
-		base.Show(isShown);
+		if (!base.Show(isShown))
+			return false;
 		if (isShown)
 			Map.EM.RemoveComponent(_connections, typeof(Frozen));
 		else
 			Map.EM.AddComponent(_connections, typeof(Frozen));
+		return true;
 	}
 
 	public void UpdateConnections()
